ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.1.0
python-multipart==0.0.6
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
scikit-learn==1.3.2
numpy==1.24.3
ultralytics==8.0.200
opencv-python==4.8.1.78
